refactor(database): condense essay helper functions into one file under app/utils
